kommentare hinzugefügt

+++ multitasking/ringbuffer.c
@@ -1,13 +1,13 @@
 #include "ringbuffer.h"
 #include "types.h"
 
-extern void printstring(char *c);
+extern void printstring(char *c); 
 
-int head = 0;
-int tail = 0;
+uint8_t head = 0;                   // Position des schreibenden Elementes
+uint8_t tail = 0;                   // Position des lesenden Elementes
 char ringbuffer[BUFFER_SIZE];
 
-int full_flag = 0;
+uint8_t full_flag = 0;
 uint8_t empty_flag = 1;
 
 int is_full(){
@@ -18,8 +18,8 @@ int is_empty(){
     return empty_flag;
 }
 
-int rb_write(char c){
-   if (is_full()){
+int rb_write(char c){      // schreibt in den Buffer und verändert head
+   if (is_full()){         // wenn der Buffer voll ist gibt er -1 zurück (als Fehlermeldung)
       return -1;
    }else{
       if (empty_flag = 1){
@@ -33,8 +33,8 @@ int rb_write(char c){
    }
 }
 
-int rb_read(char *c){
-   if (is_empty()){
+int rb_read(char *c){    // liest ein zeichen und verändert tail
+   if (is_empty()){      // wenn der Buffer leer ist gibt er -1 zurück (als Fehlermeldung)
      return -1;
    }else{
      if (full_flag = 1){
